streamlit_qa_app/requirements.txt
streamlit==1.29.0
fugashi==1.3.0
unidic-lite==1.0.8
faiss-cpu==1.7.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
transformers==4.35.2
openai==1.3.3